Updated version of merge XSL file
The original XSL file was used as the basis for this file.  Upon testing the original XSL file, not all data appeared as expected for current ELAN 4.8 and FLEx 9.0.7RC versions.  Templates were dropped and condensed into a single template (which copies from the FLEx .flextext file).  Comments were modified/added to apply to the updateds of the XSL file.