Initial commit for Python eaf tier label rename script
This script does two things.  First, update the tier labels to use less cryptic names than what is generated during the ELAN .flextext import process.  Second, it replaces the junky "A_", "B_", "C_",  etc. prefixes with the speaker's (or PARTICIPANT's) initials.  When two (or more) speakers have the same initials, a counter will be appended to the initials.  For example, the second speaker with "JS_" will be "JS2_", etc.